requirements.txt
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.7
faiss-cpu==1.11.0
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.0
fsspec==2025.5.1
hf-xet==1.1.2
huggingface-hub==0.32.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.2.1
nltk==3.9.1
numpy==2.0.2
packaging==25.0
pillow==11.2.1
PyMuPDF==1.26.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==4.1.0
setuptools==80.9.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.0+cpu --index-url https://download.pytorch.org/whl/cpu
tqdm==4.67.1
transformers==4.52.3
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
